requirements.txt
tk
opencv-python
Pillow
deepface
tf-keras